requirements.txt
pymupdf
numpy
sentence-transformers
google-generativeai
python-dotenv
streamlit